refactor: resume_yaml_to_json.rb to add image property using gravatar hash
This commit modifies the resume_yaml_to_json.rb file to add an image property to the basics section of the YAML data. If the image property is not already defined, it will be set using the gravatar hash generated from the site's email configuration.

These changes have implications for generating JSON resumes from YAML data. The addition of the image property enhances the completeness and visual appeal of the generated resumes.