feat: Add doc for function to generate Gravatar URL
This commit adds a new function `get_gravatar_url` to the `_plugins/resume_yaml_to_json.rb` file. The function takes an email address as input and returns the corresponding Gravatar URL. It also accepts an optional size parameter for customizing the image size, with a default value of 600 pixels.

The `get_gravatar_url` function can be used to easily generate Gravatar URLs for displaying user avatars or profile pictures on websites. By providing an email address, developers can retrieve the Gravatar URL associated with that email and use it in their applications.

Note: This commit also includes some minor changes in the existing code for better readability and maintainability.