feat: Add links & references section to README
This commit adds a new section called "links & references" to the README file. The section provides information about the usage of the JSON Resume schema by [jsonresume](https://jsonresume.org/).

The addition of this section enhances the documentation by providing users with relevant resources and references related to the project. It also highlights the project's compatibility with industry-standard schemas, which can be useful for developers and contributors.