feat: Template for log code to website
This commit adds the Matomo tracking code to the website. The tracking code is included in three different layouts: `default.html`, `redirect.html`, and a new file called `matm.html`. The Matomo tracking code enables cross-domain linking, tracks page views, and enables link tracking. It also sets up the tracker URL and site ID for Matomo analytics. The inclusion of this code allows for better analysis of user behavior on the website and provides valuable insights for improving its performance and user experience.

The changes can be found in the following files:
- `_includes/matm.html`
- `_layouts/default.html`
- `_layouts/redirect.html`